--- app/api/models/route.ts
@@ -1,39 +1,0 @@
-import { NextResponse } from 'next/server';
-
-export async function GET() {
-  try {
-    const response = await fetch('https://text.pollinations.ai/models', {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      // Add timeout to prevent hanging requests
-      signal: AbortSignal.timeout(10000), // 10 second timeout
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-
-    const data = await response.json();
-    return NextResponse.json(data, {
-      // Add proper headers to prevent caching issues
-      headers: {
-        'Cache-Control': 'no-cache, no-store, must-revalidate',
-        'Pragma': 'no-cache',
-        'Expires': '0',
-      },
-    });
-  } catch (error) {
-    console.error('Error fetching models:', error);
-    return NextResponse.json(
-      { error: 'Failed to fetch models' },
-      { 
-        status: 500,
-        headers: {
-          'Cache-Control': 'no-cache, no-store, must-revalidate',
-        },
-      }
-    );
-  }
-}
